bulk rename to config-foundry
