+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["uv_build>=0.8.10,<0.9.0"]
 build-backend = "uv_build"
 
 [project]
-name = "metaconf"
+name = "config-foundry"
 version = "0.4.0"
 authors = [
     {name = "Joe Marsh Rossney", email = "17361029+jmarshrossney@users.noreply.github.com"}
@@ -51,6 +51,10 @@ test = [
     "pytest>=8.4.1",
     "pytest-cov>=7.1.0",
 ]
+
+[tool.pyright]
+venvPath = "."
+venv = ".venv"
 
 [tool.ruff]
 line-length = 88
