+++ src/config_foundry/utils.py
@@ -59,7 +59,7 @@ def tree(path: str | os.PathLike) -> str:
     Constructs a tree-like representation of a directory.
 
     This is primarily for sanity-checking by comparing the output of
-    [`MetaConfig.tree`][metaconf.config.MetaConfig.tree] with an actual
+    [`MetaConfig.tree`][config_foundry.config.MetaConfig.tree] with an actual
     directory.
 
     Note:
